Add files via upload
This is created using the YOLO library. First, we need to make our data in the required YOLO format, that is, an image and labels folder where images contain several images as attached in this, and some labels like x,y, width, and height. This label is used like an annotation for an image to find the defect or detect it. This model will find any defect in an object it will find it and show it in a rectangular box to identify.